backend/requirements.txt
# SOLTEXI SAS - Backend Dependencies
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validación y configuración
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilidades
python-dotenv==1.0.0
